+++ packages/gamut/src/Layout/Column.tsx
@@ -3,7 +3,7 @@ import cx from 'classnames';
 
 import { generateResponsiveClassnames } from '../utils/generateResponsiveClassnames';
 import { ContainerElementProps, ColumnSizes, OffsetColumnSizes } from './types';
-import s from './styles/Column.module.scss';
+import styles from './styles/Column.module.scss';
 import {
   ResponsiveProperty,
   OptionalResponsiveProperty,
@@ -24,9 +24,9 @@ export const Column: React.FC<ColumnProps> = ({
   testId,
 }) => {
   const classNames = cx(
-    s.container,
+    styles.container,
     className,
-    generateResponsiveClassnames({ size, offset }, s)
+    generateResponsiveClassnames({ size, offset }, styles)
   );
   return (
     <div className={classNames} data-testid={testId}>