+++ packages/gamut/src/Markdown/libs/overrides/Table/index.tsx
@@ -1,5 +1,5 @@
 import React, { HTMLAttributes } from 'react';
-import s from './styles.module.scss';
+import styles from './styles.module.scss';
 
 export interface TableProps extends HTMLAttributes<HTMLTableElement> {
   maxHeight?: number | string;
@@ -7,7 +7,7 @@ export interface TableProps extends HTMLAttributes<HTMLTableElement> {
 
 export const Table: React.FC<TableProps> = ({ maxHeight, ...props }) => {
   return (
-    <div className={s.tableWrapper} style={{ maxHeight }}>
+    <div className={styles.tableWrapper} style={{ maxHeight }}>
       <table {...props} />
     </div>
   );