+++ packages/gamut-labs/CHANGELOG.md
@@ -1,3 +1,14 @@
+### [4.2.10](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@4.2.9...@codecademy/gamut-labs@4.2.10) (2020-10-08)
+
+
+### Other Changes
+
+* rename 's' styles to 'styles' ([#984](https://github.com/Codecademy/client-modules/issues/984)) ([0cf7210](https://github.com/Codecademy/client-modules/commit/0cf721035564f9050cdbcc9de689f149be0e96a2))
+
+### [4.2.9](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@4.2.8...@codecademy/gamut-labs@4.2.9) (2020-10-08)
+
+### [4.2.8](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@4.2.7...@codecademy/gamut-labs@4.2.8) (2020-10-08)
+
 ### [4.2.7](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@4.2.6...@codecademy/gamut-labs@4.2.7) (2020-10-06)
 
 ### [4.2.6](https://github.com/Codecademy/client-modules/compare/@codecademy/gamut-labs@4.2.5...@codecademy/gamut-labs@4.2.6) (2020-10-06)